pyproject.toml
[tool.poetry]
authors = ["Tobias Wohland <Tobias.Wohland@gmail.com>"]
description = "Logs upcoming validator duties to the console."
license = "MIT"
name = "eth-duties"
readme = "README.md"
repository = "https://github.com/TobiWo/eth-duties"
version = "0.6.0"

[tool.poetry.dependencies]
eth-typing = "==3.5.2"
fastapi = "==0.104.1"
python = "==3.12.0"
pyyaml = "==6.0.1"
requests = "==2.31.0"
sty = "==1.0.4"
uvicorn = "==0.24.0.post1"

[tool.poetry.group.dev.dependencies]
black = "==23.11.0"
isort = "==5.12.0"
mypy = "==1.7.0"
pyinstaller = "==6.2.0"
pylint = "==3.0.2"
pytest = "==7.4.4"
pytest-cov = "==4.1.0"
pytest-mock = "==3.12.0"
rope = "==1.11.0"
types-PyYAML = "==6.0.12.12"
types-requests = "==2.31.0.10"

[tool.pytest.ini_options]
pythonpath = "duties"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.isort]
profile = "black"
